refactored the code in checkout step
