+++ src/components/checkout/CheckoutSteps/PaymentStep.jsx
@@ -1,0 +1,140 @@
+export default function PaymentStep({ open, setOpen, loading, items, total }) {
+  const disabled = loading || items.length === 0;
+
+  return (
+    <div className="bg-white rounded-xl border border-gray-200 shadow-sm overflow-hidden">
+      <div
+        className="px-4 py-3 cursor-pointer flex items-center justify-between border-b border-gray-200 bg-white"
+        onClick={() => setOpen(!open)}
+      >
+        <div className="flex items-center space-x-3">
+          <div className="w-8 h-8 bg-gray-100 rounded-lg flex items-center justify-center">
+            <svg
+              className="w-5 h-5 text-gray-700"
+              fill="none"
+              stroke="currentColor"
+              viewBox="0 0 24 24"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                strokeWidth={2}
+                d="M3 10h18M7 15h1m4 0h1m-7 4h12a3 3 0 003-3V8a3 3 0 00-3-3H6a3 3 0 00-3 3v8a3 3 0 003 3z"
+              />
+            </svg>
+          </div>
+          <div>
+            <h3 className="text-lg font-bold text-gray-900">Payment</h3>
+            <p className="text-sm text-gray-500">
+              Review your total and pay securely
+            </p>
+          </div>
+        </div>
+        <button
+          type="button"
+          aria-label="Toggle payment"
+          className={`transition-transform ${open ? "rotate-180" : "rotate-0"}`}
+        >
+          <svg
+            className="w-8 h-8 text-gray-600"
+            viewBox="0 0 20 20"
+            fill="currentColor"
+          >
+            <path
+              fillRule="evenodd"
+              d="M5.23 7.21a.75.75 0 011.06.02L10 10.94l3.71-3.71a.75.75 0 111.06 1.06l-4.24 4.24a.75.75 0 01-1.06 0L5.21 8.29a.75.75 0 01.02-1.08z"
+              clipRule="evenodd"
+            />
+          </svg>
+        </button>
+      </div>
+      {open && (
+        <div className="p-6 md:w-1/2 mx-auto">
+          <div className="flex items-center justify-between p-4 bg-gray-50 rounded-lg border border-gray-200 mb-4">
+            <div className="flex items-center space-x-3">
+              <img
+                src="/stripe.png"
+                className="w-8 h-8 rounded-full"
+                alt="Stripe"
+              />
+              <div>
+                <p className="font-semibold text-gray-900">Secure Payment</p>
+                <p className="text-sm text-gray-600">Powered by Stripe</p>
+              </div>
+            </div>
+            <svg
+              className="w-5 h-5 text-green-500"
+              fill="none"
+              stroke="currentColor"
+              viewBox="0 0 24 24"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                strokeWidth={2}
+                d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z"
+              />
+            </svg>
+          </div>
+          <button
+            type="submit"
+            disabled={disabled}
+            className={`w-full py-4 px-6 rounded-lg font-bold text-white transition-all duration-200 flex items-center justify-center space-x-2 ${
+              disabled
+                ? "bg-gray-400 cursor-not-allowed"
+                : "bg-primary hover:opacity-90 shadow-lg hover:shadow-xl"
+            }`}
+          >
+            {loading ? (
+              <>
+                <svg
+                  className="animate-spin w-5 h-5"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                >
+                  <circle
+                    className="opacity-25"
+                    cx="12"
+                    cy="12"
+                    r="10"
+                    stroke="currentColor"
+                    strokeWidth="4"
+                  ></circle>
+                  <path
+                    className="opacity-75"
+                    fill="currentColor"
+                    d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
+                  ></path>
+                </svg>
+                <span>Processing...</span>
+              </>
+            ) : (
+              <>
+                <svg
+                  className="w-5 h-5"
+                  fill="none"
+                  stroke="currentColor"
+                  viewBox="0 0 24 24"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth={2}
+                    d="M3 10h18M7 15h1m4 0h1m-7 4h12a3 3 0 003-3V8a3 3 0 00-3-3H6a3 3 0 00-3 3v8a3 3 0 003 3z"
+                  />
+                </svg>
+                <span>
+                  Pay $
+                  {total.toLocaleString("en-US", {
+                    minimumFractionDigits: 2,
+                    maximumFractionDigits: 2,
+                  })}
+                </span>
+              </>
+            )}
+          </button>
+        </div>
+      )}
+    </div>
+  );
+}
